+++ module/Task/src/Domain/Task/Event/TaskCompleted.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Task\Domain\Task\Event;
+
+use DateTimeImmutable;
+use ExtendsSoftware\ExaPHPExample\Shared\Domain\DomainEventInterface;
+use ExtendsSoftware\ExaPHPExample\Task\Domain\Task\TaskId;
+
+final readonly class TaskCompleted implements DomainEventInterface
+{
+    public function __construct(
+        private(set) TaskId $taskId,
+        private(set) DateTimeImmutable $completedAt,
+    ) {}
+}